fix(frontend): resolve codebook manual entry routing and layout bugs
- Merged codebook.py into codebooks.py to fix routing conflicts
- Moved templates from codebook to codebooks directory
- Fixed HTML layout bug in preview.html that caused forms to close early
- Added 50-theme Javascript limit to preview.html
- Updated success.html route reference to avoid 500 error
