--- Frontend/web/templates/codebook/list.html
@@ -1,232 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Codebooks{% endblock %}
-
-{% block head_extra %}
-<link href="https://fonts.googleapis.com/css2?family=Outfit:wght@300;400;600;700&display=swap" rel="stylesheet">
-<style>
-  :root {
-    --primary-color: #4f46e5;
-    --primary-hover: #4338ca;
-    --border-radius: 12px;
-    --card-shadow: 0 10px 25px -5px rgba(0, 0, 0, 0.05), 0 8px 10px -6px rgba(0, 0, 0, 0.05);
-  }
-
-  .list-container {
-    font-family: 'Outfit', 'Inter', sans-serif;
-    max-width: 900px;
-    margin: 2rem auto;
-    padding: 0 1.5rem;
-  }
-
-  .page-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 2rem;
-  }
-
-  .page-title {
-    font-size: 2rem;
-    font-weight: 700;
-    color: #1e293b;
-    margin: 0;
-  }
-
-  .page-subtitle {
-    font-size: 1rem;
-    color: #64748b;
-    margin: 0.25rem 0 0;
-  }
-
-  .btn {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.4rem;
-    padding: 0.7rem 1.4rem;
-    font-weight: 600;
-    border-radius: 8px;
-    cursor: pointer;
-    border: none;
-    transition: all 0.2s ease;
-    text-decoration: none;
-    font-size: 0.95rem;
-  }
-
-  .btn-primary {
-    background: var(--primary-color);
-    color: #ffffff;
-  }
-
-  .btn-primary:hover {
-    background: var(--primary-hover);
-  }
-
-  .codebook-grid {
-    display: grid;
-    gap: 1.25rem;
-  }
-
-  .codebook-card {
-    background: #ffffff;
-    border: 1px solid #e2e8f0;
-    border-radius: var(--border-radius);
-    padding: 1.5rem 2rem;
-    box-shadow: var(--card-shadow);
-    display: grid;
-    grid-template-columns: 1fr auto;
-    align-items: center;
-    gap: 1rem;
-    transition: all 0.25s cubic-bezier(0.4, 0, 0.2, 1);
-    text-decoration: none;
-    color: inherit;
-    position: relative;
-    overflow: hidden;
-  }
-
-  .codebook-card::before {
-    content: '';
-    position: absolute;
-    top: 0; left: 0;
-    width: 4px; height: 100%;
-    background: linear-gradient(180deg, #4f46e5 0%, #7c3aed 100%);
-    opacity: 0;
-    transition: opacity 0.2s ease;
-  }
-
-  .codebook-card:hover {
-    border-color: #c7d2fe;
-    transform: translateX(4px);
-    box-shadow: 0 20px 25px -5px rgba(79, 70, 229, 0.06);
-  }
-
-  .codebook-card:hover::before {
-    opacity: 1;
-  }
-
-  .codebook-name {
-    font-size: 1.15rem;
-    font-weight: 600;
-    color: #1e293b;
-    margin: 0 0 0.3rem;
-  }
-
-  .codebook-meta {
-    font-size: 0.85rem;
-    color: #94a3b8;
-    display: flex;
-    gap: 1rem;
-    flex-wrap: wrap;
-  }
-
-  .meta-badge {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.25rem;
-    background: #f1f5f9;
-    border-radius: 4px;
-    padding: 0.2rem 0.55rem;
-    font-weight: 500;
-    color: #475569;
-  }
-
-  .version-badge {
-    background: #f5f3ff;
-    color: var(--primary-color);
-    font-weight: 700;
-  }
-
-  .arrow-icon {
-    color: #cbd5e1;
-    font-size: 1.2rem;
-    transition: color 0.2s ease;
-  }
-
-  .codebook-card:hover .arrow-icon {
-    color: var(--primary-color);
-  }
-
-  /* Empty state */
-  .empty-state {
-    text-align: center;
-    padding: 5rem 2rem;
-    background: #f8fafc;
-    border: 2px dashed #e2e8f0;
-    border-radius: var(--border-radius);
-  }
-
-  .empty-icon {
-    font-size: 4rem;
-    margin-bottom: 1rem;
-    opacity: 0.6;
-  }
-
-  .empty-title {
-    font-size: 1.4rem;
-    font-weight: 600;
-    color: #475569;
-    margin-bottom: 0.5rem;
-  }
-
-  .empty-desc {
-    color: #94a3b8;
-    margin-bottom: 2rem;
-  }
-
-  .custom-error-alert {
-    background: #fef2f2;
-    border-left: 4px solid #ef4444;
-    color: #b91c1c;
-    padding: 1rem 1.25rem;
-    border-radius: 6px;
-    margin-bottom: 2rem;
-  }
-</style>
-{% endblock %}
-
-{% block content %}
-<div class="list-container">
-  <div class="page-header">
-    <div>
-      <h1 class="page-title">Codebooks</h1>
-      <p class="page-subtitle">Your saved thematic frameworks, ready to guide analysis runs.</p>
-    </div>
-    <a href="{{ url_for('codebook.upload_form') }}" class="btn btn-primary" id="new-codebook-btn">
-      + New Codebook
-    </a>
-  </div>
-
-  {% if error %}
-    <div class="custom-error-alert">
-      <strong>Error:</strong> {{ error }}
-    </div>
-  {% endif %}
-
-  {% if codebooks %}
-    <div class="codebook-grid" id="codebook-list">
-      {% for cb in codebooks %}
-        <a href="{{ url_for('codebook.success', codebook_id=cb.id) }}"
-           class="codebook-card"
-           id="codebook-card-{{ loop.index }}">
-          <div>
-            <p class="codebook-name">{{ cb.name }}</p>
-            <div class="codebook-meta">
-              <span class="meta-badge version-badge">v{{ cb.version }}</span>
-              <span class="meta-badge">Project: {{ cb.project_id }}</span>
-              <span class="meta-badge">By: {{ cb.created_by }}</span>
-            </div>
-          </div>
-          <span class="arrow-icon">→</span>
-        </a>
-      {% endfor %}
-    </div>
-  {% else %}
-    <div class="empty-state" id="empty-codebooks">
-      <div class="empty-icon">📋</div>
-      <div class="empty-title">No codebooks yet</div>
-      <p class="empty-desc">Upload a CSV or enter themes manually to create your first codebook.</p>
-      <a href="{{ url_for('codebook.upload_form') }}" class="btn btn-primary">Create Your First Codebook</a>
-    </div>
-  {% endif %}
-</div>
-{% endblock %}
