+++ Frontend/web/templates/codebooks/preview.html
@@ -188,6 +188,7 @@
 
 {% block content %}
 <div class="preview-container">
+  <form id="codebook-form" method="post" action="{{ url_for('codebooks.confirm_submit') }}">
   <div class="header-card">
     <h1>Preview &amp; Confirm Themes</h1>
     <p style="color: #64748b; margin-top: -1rem; margin-bottom: 1.5rem;">Verify, modify, or add themes to your codebook before locking it in.</p>
@@ -198,7 +199,6 @@
       </div>
     {% endif %}
 
-    <form id="codebook-form" method="post" action="{{ url_for('codebook.confirm_submit') }}">
       <div class="form-group">
         <label for="codebook_name" class="theme-label">Codebook Name</label>
         <input type="text"
@@ -248,6 +248,11 @@
 
 <script>
   function addRow() {
+    const rows = document.querySelectorAll('.theme-row');
+    if (rows.length >= 50) {
+      alert("A codebook cannot contain more than 50 themes.");
+      return;
+    }
     const container = document.getElementById('themes-list-container');
     const newRow = document.createElement('div');
     newRow.className = 'theme-row';
